Update Plugins to Classes
