+++ src/commands/utility/Plugins.ts
@@ -1,34 +1,44 @@
 import Discord from "discord.js";
 import packageJson from "../../../package.json" assert { type: 'json' };
 import { SHAII_LOGO } from "../../constants";
-import Naoko from "../../naoko/Naoko";
 import { CommandExecuteResponse } from "../../types";
 import AbstractCommand, { CommandData } from '../AbstractCommand';
 import command from '../../decorators/command';
-
-const PLUGIN_ICON = "https://cdn.discordapp.com/attachments/911762334979084368/923234119553519636/unknown.png";
+import { PluginManager } from "../../plugins/PluginManager";
+import { delay, inject, injectable } from "@triptyk/tsyringe";
 
 @command()
+@injectable()
 class Plugins extends AbstractCommand {
+  private static PLUGIN_ICON = "https://cdn.discordapp.com/attachments/911762334979084368/923234119553519636/unknown.png";
+
+  constructor(
+    @inject(delay(() => PluginManager)) private pluginManager: PluginManager,
+  ) {
+    super();
+  }
+
   execute(): CommandExecuteResponse | Promise<CommandExecuteResponse> {
+    const plugins = this.pluginManager.getAll();
+
     const embed = new Discord.EmbedBuilder()
       .setAuthor({ name: `Naoko v${packageJson.version}`, iconURL: SHAII_LOGO })
-      .setThumbnail(PLUGIN_ICON)
+      .setThumbnail(Plugins.PLUGIN_ICON)
       .setColor("#FF00B6")
       .addFields({
         inline: true,
         name: "Plugins",
-        value: Naoko.plugins.map((plugin) => `🔌 ${plugin.name}`).join("\n"),
+        value: plugins.map((plugin) => `🔌 ${plugin.pluginData.name}`).join("\n"),
       })
       .addFields({
         inline: true,
         name: "Version",
-        value: Naoko.plugins.map((plugin) => plugin.version).join("\n"),
+        value: plugins.map((plugin) => plugin.pluginData.version).join("\n"),
       })
       .addFields({
         inline: true,
         name: "Enabled?",
-        value: Naoko.plugins.map((plugin) => plugin.state === 1 ? '✅' : '❌').join("\n"),
+        value: plugins.map((plugin) => plugin.pluginData.enabled ? '✅' : '❌').join("\n"),
       });
 
     return { embeds: [embed] };
