+++ src/decorators/command.ts
@@ -1,11 +1,13 @@
-import { container } from '@triptyk/tsyringe';
+import { container, injectable } from '@triptyk/tsyringe';
+import AbstractCommand from 'src/commands/AbstractCommand';
 
-type constructor<T> = {
+type constructor<T = AbstractCommand> = {
   new (...args: any[]): T;
 }
 
 export default function command<T>(): (target: constructor<T>) => void {
   return function(target: constructor<T>): void {
+    injectable()(target);
     container.registerSingleton(target);
     container.register('naokoCommand', target);
   };
