--- src/plugins/baking-const/baking.ts
@@ -1,15 +1,0 @@
-export const CAKES = [
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236192646320148/cake_1.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236192843464754/cake_2.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236193032183898/pink-cake.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236193258692758/lie.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236193455820851/cake.jpg",
-];
-
-export const PIES = [
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236236317417512/pie-1.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236236631998495/pies.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236236908798033/coco.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236237252743198/apple.jpg",
-  "https://cdn.discordapp.com/attachments/634839969822801998/964236237554745425/beryy.jpeg",
-];
